add swaps hetper and fix selectet items in create offer
